Tweaks behavior of client side session warning message popup CS-4457
Now handles n depth frames (which becomes important for new message
sender, which has iframes in iframes)
Bubbles up renew requests to the top possible window which could be
responsible for handling the session expiration warning.
This also handles cases where there is a same-origin restriction in place
for third parties (PowerSchool) in which the "top" window responsible is
actually a child.