Attempts to validate that content is useable
At a minimum, we have to be sure no data is allowed through which could
corrupt the database. This validator wasn't even trying to make sure the
data passed in wasn't garbage! Now it checks against allowed
content ids stored in session data, but could be made much much better.