Enhance method and simplify logic when jobs are involved, CS-4623
If we get to the point of checking a message group's job, just fall through to the
part of code which already checks jobs and allow it to handle the logic for us.
A user can always see message groups for jobs which they can also see, do it
doesn't make sense to have two different groups of job checking logic in the
same method.
Also, move the monitor check logic out to make it easier to follow. It's also
re-useable now as well.