Corrects invalid query, CS-5992
Query result was concatenating student pkey values into the query string directly w/o escaping them.
It also attempted to compare them with the person id object, which is the auto-increment database
id and not the pkey for the person object.

The above caused database errors to be logged and created a vector for sql injection attacks through
the student pkey value.