CS-7275-SMSReport fix bug in form handling.
You can't have sticky elements in forms that are built into the
form definition itself. When we change the SMS phone in the search
box, the form thinks the data entity has changed out from under it.
These forms are primarily designed for CRUD operations, it seems.

Also applied a request to change the column order and heading.