CS-7275 fix VIEW to be relative to current customer db, fix CSV headers, fix button order.
We had accidentally made the VIEW definition always reference customer database `c_1`.

The CSV headers were in the wrong order relative to the columns,
also need to change "Person Pkey" to "Unique ID"

Make the "Search for Single SMS" button first, which allows someone to simply hit return
to search for a number.