Updates numerous queries pulling reports from reportcontact to ignore records with "declined" result CS-7440 CS-7451
The "declined" state is added for reportcontact records of contact destinations that the person has opted out of (declined). This record results in no transmission, but serves as a permanent "admission ticket" that ICRA will use to allow ICRA users connected with those recipients to see those messages. Otherwise if they were opted out of everything, it would be impossible for ICRA to allow the user to see anything at all. Since the "declined" records being added represent net new data in the system, we are filtering ALL of them out of preexisting queries within Kona so that the new data does not impact report results.