Changes SDD send email form to pre-populate ALL form options with override capability to let user just hit SEND if they like it
It is painful filling out all the fields on the form especially when some of them may well already be known from the context of the logged in user, etc. This simply pre-polulates some values so that they can just submit if they don't care so much...

Also applies some refactoring for improved code efficiency in this area.