Fixes bug dealing with password protected SDD portion delivery CS-7552
Problem came from recent refactoring of SDD delivery code when adding support for GET retrieval of email attachments for ICRA delivery. The refactoring was not completed and left a large, old, duplicate chunk of code behind as a point of reference for the original implementation just to be able to correct a conditional check (for $verify) in the new implementation, now in apiGetDocument.php. The result was that the $verify value was not being set correctly and the password was always getting set to null when interfacing with the CommSuite Thrift API even though the form POSTed with a proper password. Now these factors have been corrected and password-protected documents are delivering correctly.