Adds a CKEditor plugin to handle image resizing for webkit-based browsers CS-7254 CS-8151
Note that the plugin comes from https://github.com/sstur/ck-dragresize, and this commit also adds a minor JS hack to make its working mechanism's output consistent with the non-webkit browsers' result (which is to use html native img tag attributes for width and height instead of style attribute width/height properties. Although these have the same visible result in the client, on the server side we are using regular expressions to parse the image tag and the width/height attributes are already accounted for.