Introduces a simple, custom JS object "Node Registry" utility script CS-7254 CS-8190
The basic idea is to be able to take any given JS object and wrap it as a "Node" which gets added to the registry. Each Node can have named events added to it and then any other JS code can easily invoke events against any of the registered nodes without having to directly link those normally disparate pieces of JS together.