Fixes an issue with IE11 not capturing image resize changes CS-7254 CS-8242
Problem was caused by something *really* strange with IE11 doing the resize operation, capturing the original image dimensions in height/width HTML attributes, and the resize override dimensions in style attribute's height/width properties to override them. The server-side code was set up to ONLY accept resize dimensions from the height/width img tag attributes, so it appeared to be unchanged. Preview was still working because the image tag was preserved with the style attribute as it was found in the editor and as modified by IE11.

Solution was to make a more flexible chunk of code which would give preference to style attribute properties in either order (regex-based matching, so ordering matters; e.g. height then width vs width then height) and as a last resort checks height/width image tag attribute values. In this way each mode of image resizing is accomodated in its native form without hacking.