Corrects support for PNG type images CS-7254 CS-8069
PNG compression is 0-9 range, not 0-99, so the value "90" copy/pasted from the originating JPEG handling code was out of range which caused resize operations to fail and produce a broken image reference. Also PHP's documentation for getimagesize() function is incorrect on the type constants needed to identify the mimetypes for PNG/JPG/GIF, etc, so these are corrected to match the GDlib documentation.