Alters Start Date validator to account for optional facebook page selection CS-7441 CS-8338
Originally if a job's message had facebook content in it, the user was required to have a valid FB auth token so that the start date selected could be compared to the expiration date of the FB token. But now that FB page selection is optional, we only take the FB token expiration into consideration for start date if a page has been selected - if no page has been selected, then the token will not be used and therefore the expiration date does not matter. Hence it is no longer permitted to interfere with selection of furutre start dates beyond the expiration of the current FB token.