Adds one more condition to the HostedAttachment report query UI-101 UI-246
There were different restrictions on the query to get the list of jobs than on the query to get the details of the job on the following page. This was revealed by a separate bug wherein the message data is not getting created correctly at the time of message creation and/or job send - there was enough data to get the job into the list, but not enough data to satisfy the details page that followed. This will be separately addressed when we have steps to reproduce in a follow-up bug report.