Stops a bunch of reporting pages from generating useless errors UI-198 UI-263
"aspcalls" needs to be configured in settings for any of these reports to work. Without the setting the UI still lets a user click aon a dozen links to pull reports that cannot work and throws errors. now the error logging is silenced with a note to the user that it isn't configured.