Makes last argument optional to suppress error message UI-232 UI-314
The fourth argument not supplied in a call generates a warning but still runs the function with the 4th argument as null. Since it is supposed to be a boolean, null evaluates to false, so we default the value to false to prevent any change in behavior and make the error go away.