Removes an attempt to set time_zone on non-existent MySQL connection UI-229 UI-305
Tracing the code leading up to this, it doesn't appear that it could ever be successful since there is no attempt to connect to a database before running the quickupdate to set the timezone for the database session. Since it cannot have worked, then it should be harmless to remove. For now it is commented out until this can be verified; later we can remove it permanently.