Applies proper Thrift namespace for TException to enable retry on error UI-217 UI-295
PHP Does not throw an error when you attempt to catch an undefined exception class name, so although we have been using "TException" in our catch statements which appears legitimate to the human reader, it has actually been an undefined class because there was no "use" statement to bring TException out of the Thrift namespace and into the default namespace. This has prevented these exceptions from getting caught and defeated the associated retry logic to aid in self-recovery on transient errors.