Use GET to fetch a HA document if possible UI-271
This modifies messagelink/js/sdd.js to use a GET request to fetch
a document if a password is not required. This fixes playback
of audio files on mobile browsers, which will re-request (as
a GET) any audio files in order to play them back.

The bug fixed by this commit will still affect password-protected
downloads