Fixes issue with conflicting consent data for duplicate numbers UI-1224 UI-1227
The duplicate numbers would cause the edit form to display a select field for the first entry (of the duplicate entries), but static text for the second. Then when the form was submitted, the submit handler would see both and request the value for each, the first of which would have the newly selected value in it, but the second of which would have the original value, overriding the newly selected value. This would give the effect that the saved value to GDR would appear to be unchanged. The solution was to prevent the submit handler from seeing the duplicate entries once the original entry's selection value was collected up.