Fix for incomplete sms status report CS-10029
SMS status report query was grouping by organization, resulting in one report row per organization found in pool of otherwise eligible rows.

Group by sms number is the way the query used to be structured and that approach has been reinstated.

We are not completely sure that this won't have some ill effect for other cases, but the current report result is not usable for customers.